debug /ctrl and /sys, lots lots emojis :grin:
++ added a emoji dict with 845 emojis depending on
https://github.com/github/gemoji/blob/master/db/emoji.json
A list of all emojis can be found here:
http://apps.timwhitlock.info/emoji/tables/unicode

use the text from 'description' column (on the page mentioned above) as
<iconDescr> to get the desired emoji.
in python: TelegramPlugin.gEmo('<iconDescr>')
in messages format string: {emo:<iconDescr>}
a few shortcuts for easier use  can be found in TelegramPlugin.emojis{}

+- updated emoji format handling in TMSG._sendNotification()
added EmojyFormatter() class to do that

+- debugged /ctrl and /sys, updated to safer hashes
the /ctrl and  /sys commands now use hashes of the command names as
parameter.
so uncomon chars in command names will not corrupt command link like
/sys_+_up
(the '+' ends the link, so klicking it will only execute /sys_)
++ added simple but good enogh hash function for command hashing
++ added "path" to controlnames in get_controls_recursively()
otherwise there may controls with same names when there are two
containers with
one 'on' button in each. Without appending container name, the name
wont be unique and produces same hashes. Also have to append the command
for
hash because controls with same names are allowed in one container. So
as long as no
ctrls with same name and same command in same container we are safe
:cool: