Fix for issue #44
Only attempt to access such keys in the provided settings that
are actually there and in the frontend manually persist chats
as received from own plugin endpoint to global settings
endpoint in order to trigger on_settings_save processing.

This is a bit dirty but basically emulates what the click on
Save in the settings dialog used to do (I think - I have to admit
I was a bit confused by the whole back-and-forth). After completion
of the new save call it's actually no longer necessary to explicitly
click Save on the Settings dialog, but it was completely unclear
to me how to properly set the new status for a newly added chat
without that roundtrip.